Reorganize the classes a bit
