+++ src/main/kotlin/dev/monosoul/jooq/JooqExtension.kt
@@ -1,8 +1,8 @@
 package dev.monosoul.jooq
 
-import dev.monosoul.jooq.JooqDockerPluginSettings.Database
-import dev.monosoul.jooq.JooqDockerPluginSettings.WithContainer
-import dev.monosoul.jooq.JooqDockerPluginSettings.WithoutContainer
+import dev.monosoul.jooq.settings.JooqDockerPluginSettings
+import dev.monosoul.jooq.settings.JooqDockerPluginSettings.WithContainer
+import dev.monosoul.jooq.settings.JooqDockerPluginSettings.WithoutContainer
 import org.gradle.api.Action
 import java.io.Serializable
 
@@ -18,28 +18,6 @@ open class JooqExtension : Serializable {
     fun withoutContainer(configure: Action<WithoutContainer>) {
         pluginSettings = WithoutContainer.new {
             configure.execute(this)
-        }
-    }
-
-    class Jdbc : Serializable {
-        var schema = "jdbc:postgresql"
-        var driverClassName = "org.postgresql.Driver"
-        var urlQueryParams = ""
-    }
-
-    class Image(db: Database.Internal) : Serializable {
-        var repository = "postgres"
-        var tag = "11.2-alpine"
-        var envVars: Map<String, Any> = mapOf(
-            "POSTGRES_USER" to db.username,
-            "POSTGRES_PASSWORD" to db.password,
-            "POSTGRES_DB" to db.name
-        )
-        var testQuery = "SELECT 1"
-        var command: String? = null
-
-        internal fun getImageName(): String {
-            return "$repository:$tag"
         }
     }
 }
